requirements.txt
pytest
numpy
scipy
matplotlib
sklearn.metrics
